Removes unnecessary conditional statement, #20
CCN is reduced to 20.

The change is valid, since, first of all, len(lines) is obviously
greater than 0 (append few lines above), and, secondly, the loop on
line 94 can be only entered if loop on line 91 was entered.
`c.function` always returns non-empty string as its last line,
so the second check is always true too.